external validation less strict

+++ externalN.py
@@ -14,6 +14,7 @@ parser.add_argument('model_name', type=str, help="name of trained model")
 parser.add_argument('patience', type=int, help="number of validations without improvement before training is stopped")
 parser.add_argument('history', type=str, help="file used to store validation history")
 parser.add_argument('num_translations', type=int, help="number of translations for each input")
+parser.add_argument('-r', '--retries', dest='retries', type=int, default=1, help="number of times to restart patience counting if patience expired with 0 successes (default: \'%(default)s\')")
 parser.add_argument('-ll', '--llvm', dest='l', help="generate LLVM code", action='count')
 parser.add_argument('-pt', '--parse-tree', dest='p', help="generate parse tree code", action='count')
 
@@ -27,6 +28,7 @@ v = args.dataset
 mdir = args.model_directory
 m = args.model_name
 p = args.patience
+r = args.retries
 h = args.history
 k = args.num_translations
 
@@ -59,6 +61,7 @@ if vals:
 	hf = vals[3]
 	ht = vals[4]
 	count = vals[5]
+	retry = vals[6]
 	better = False
 	if (ns+ni) > (hs+hi):
 		better = True
@@ -76,25 +79,31 @@ if vals:
 	else:
 		count += 1
 		if count > p:
-			print 'No progress for last '+str(p)+' validations. Terminating!'
-			for proc in psutil.process_iter():
-				if proc.name() == 'python':
-					if proc.pid != os.getpid():
-						u = proc.uids()
-						if os.getuid() in [u.real, u.effective, u.saved]:
-							print '\t Kiliing process '+str(proc.pid)
-							proc.kill()
-							os.remove(h)
-							cleanup()
-							sys.exit(0)
+			retry += 1
+			if retry == r:
+				print 'No progress for last '+str(p)+' validations ('+str(retry)+' retries). Terminating!'
+				for proc in psutil.process_iter():
+					if proc.name() == 'python':
+						if proc.pid != os.getpid():
+							u = proc.uids()
+							if os.getuid() in [u.real, u.effective, u.saved]:
+								print '\t Kiliing process '+str(proc.pid)
+								proc.kill()
+								os.remove(h)
+								cleanup()
+								sys.exit(0)
+			else:
+				print 'No progress for last '+str(p)+' validations but still no successes. Retrying. ('+str(retry)+'/'+str(r)+' retries)'
+				with open(h,'w') as history:
+					history.write(str(hi)+'\t'+str(hs)+'\t'+str(hp)+'\t'+str(hf)+'\t'+str(ht)+'\t0\t'+str(retry)+'\n')
 		else:
 			with open(h,'w') as history:
-				history.write(str(hi)+'\t'+str(hs)+'\t'+str(hp)+'\t'+str(hf)+'\t'+str(ht)+'\t'+str(count)+'\n')
+				history.write(str(hi)+'\t'+str(hs)+'\t'+str(hp)+'\t'+str(hf)+'\t'+str(ht)+'\t'+str(count)+'\t'+str(retry)+'\n')
 else:
 	with open(h,'w') as history:
 		shutil.copy(os.path.join(mdir,m+'.npz.dev.npz'),os.path.join(mdir,m+'.npz.best.npz'))
 		shutil.copy(os.path.join(mdir,m+'.npz.dev.npz.json'),os.path.join(mdir,m+'.npz.best.npz.json'))
-		history.write(str(ni)+'\t'+str(ns)+'\t'+str(np)+'\t'+str(nf)+'\t'+str(nt)+'\t0\n')
+		history.write(str(ni)+'\t'+str(ns)+'\t'+str(np)+'\t'+str(nf)+'\t'+str(nt)+'\t0\t0\n')
 
 os.remove(v+'.corpus.'+str(k)+'.out')
 os.remove(v+'.'+str(k)+'.alignment')
